requirements.txt
langchain
python-dotenv
langchain-openai 
langchain-core
langchain-community
bs4
faiss-cpu
pypdf
arxiv
pymupdf
wikipedia
lxml
langchain_huggingface
langchain_groq
langgraph
langgraph-cli[inmem]
langchain-xai
langchainhub
streamlit